Add error format (simple), can be configured from configuration.
Specifying log-access-format/log-error-format in configuration defines if logs should be handled or not, and the associated format.
Access/error logging is disabled by default.
Error log format seems not standardised as common or combined log format.

For now, it is only possible to log on STDERR, but it should be possible to log in files or other output.